+++ packages/gatsby-plugin-theme-ui/gatsby-ssr.js
@@ -2,6 +2,13 @@ import { jsx, InitializeColorMode } from 'theme-ui'
 
 export { wrapRootElement } from './src/provider'
 
-export const onRenderBody = ({ setPreBodyComponents }) => {
-  setPreBodyComponents([jsx(InitializeColorMode, { key: 'theme-ui-no-flash' })])
+export const onRenderBody = (
+  { setPreBodyComponents },
+  { noFlashScript = true }
+) => {
+  if (noFlashScript) {
+    setPreBodyComponents([
+      jsx(InitializeColorMode, { key: 'theme-ui-no-flash' }),
+    ])
+  }
 }